@@ -10,5 +10,5 @@ import (
 	_ "github.com/vmware-tanzu/carvel-ytt/cmd/ytt"
 	_ "golang.org/x/tools/cmd/goimports"
 	_ "sigs.k8s.io/controller-tools/cmd/controller-gen"
-	_ "sigs.k8s.io/kustomize/kustomize/v4"
+	_ "sigs.k8s.io/kustomize/kustomize/v5"
 )
