@@ -1,5 +1,5 @@
 /*
-Copyright 2020 VMware Inc.
+Copyright 2020-2023 VMware Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -22,9 +22,9 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/vmware-labs/reconciler-runtime/apis"
 	rtesting "github.com/vmware-labs/reconciler-runtime/testing"
-	"github.com/vmware-labs/reconciler-runtime/validation"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/validation/field"
 )
 
 func TestPodIntentDefault(t *testing.T) {
@@ -57,11 +57,11 @@ func TestPodIntentValidate(t *testing.T) {
 	for _, c := range []struct {
 		name     string
 		target   *PodIntent
-		expected validation.FieldErrors
+		expected field.ErrorList
 	}{{
 		name:     "empty",
 		target:   &PodIntent{},
-		expected: validation.FieldErrors{},
+		expected: field.ErrorList{},
 	}, {
 		name: "empty image pull secret",
 		target: &PodIntent{
@@ -69,22 +69,24 @@ func TestPodIntentValidate(t *testing.T) {
 				ImagePullSecrets: []corev1.LocalObjectReference{{}},
 			},
 		},
-		expected: validation.ErrMissingField("spec.imagePullSecrets[0].name"),
+		expected: field.ErrorList{
+			field.Required(field.NewPath("spec", "imagePullSecrets").Index(0).Child("name"), ""),
+		},
 	}} {
 		t.Run(c.name, func(t *testing.T) {
-			actual := c.target.Validate()
+			actual := c.target.validate()
 			if diff := cmp.Diff(c.expected, actual); diff != "" {
 				t.Errorf("Validate() (-expected, +actual) = %v", diff)
 			}
-			create := c.target.ValidateCreate()
+			_, create := c.target.ValidateCreate()
 			if diff := cmp.Diff(c.expected.ToAggregate(), create); diff != "" {
 				t.Errorf("ValidateCreate() (-expected, +actual) = %v", diff)
 			}
-			update := c.target.ValidateUpdate(nil)
+			_, update := c.target.ValidateUpdate(nil)
 			if diff := cmp.Diff(c.expected.ToAggregate(), update); diff != "" {
 				t.Errorf("ValidateUpdate() (-expected, +actual) = %v", diff)
 			}
-			delete := c.target.ValidateDelete()
+			_, delete := c.target.ValidateDelete()
 			if diff := cmp.Diff(nil, delete); diff != "" {
 				t.Errorf("ValidateUpdate() (-expected, +actual) = %v", diff)
 			}
